requirements.txt
aiogram>=3.5
SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic-settings>=2.2
APScheduler>=3.10
python-dotenv>=1.0
google-generativeai>=0.3.0

